requirements.txt
streamlit>=1.28.0
tensorflow>=2.13.0
numpy>=1.24.0
google-generativeai>=0.3.0
keras>=2.13.0

